Major 2.0 refactor first pass
* refactor commands into individual files
* drop support for hub
* clenaup arg parsing
